remove required session in app
